- Summary: CheckIn: Bugzilla bug #206707--Double type is shown in Output Columns tab [07]
- Bug (s) Resolved:
[Bugzilla] [#206707]--Double type is shown in Output Columns tab [07]

- Description:
In current implementation the data type name in data set view is based on the data type name mapping of its data type. For the original data type is float or double, it will converted to data type to DataType.DOUBLE_TYPE in adaptModelDataType method of class DataAdapterUtil. This type will be converted to Double for type name in current mapping which based on the DataType class in package org.eclipse.birt.core.data.

Fixed the names array element from Double to Float in DataType class. Thus it will be consistent with the computedcolumn definition in data types.

This modify need update 3 txt files for test case can run normally:
 /org.eclipse.birt.data.tests/test/org/eclipse/birt/data/engine/impl/golden/ResultMetaDataTest.test1_SelectAll.txt,
/org.eclipse.birt.data.tests/test/org/eclipse/birt/data/engine/impl/golden/ResultMetaDataTest.test2_ResultHints.txt,
/org.eclipse.birt.data.tests/test/org/eclipse/birt/data/engine/impl/golden/ResultMetaDataTest.test4_ProjectedCol.txt


- Code Reviewer:
Lin Zhu/Mingxia Wu

- Tests Description:
Manual Test


- Files Edited:
"/org.eclipse.birt.core/src/org/eclipse/birt/core/data/DataType.java"
"/org.eclipse.birt.data.tests/test/org/eclipse/birt/data/engine/impl/golden/ResultMetaDataTest.test1_SelectAll.txt"
"/org.eclipse.birt.data.tests/test/org/eclipse/birt/data/engine/impl/golden/ResultMetaDataTest.test2_ResultHints.txt"
"/org.eclipse.birt.data.tests/test/org/eclipse/birt/data/engine/impl/golden/ResultMetaDataTest.test4_ProjectedCol.txt"


- Files Added:

- Files Deleted:

- Notes to Build Team:
N/A

- Notes to Developers:
N/A

- Notes to QA:
N/A

- Notes to Documentation:
N/A

+++ core/org.eclipse.birt.core/src/org/eclipse/birt/core/data/DataType.java
@@ -40,7 +40,7 @@ public abstract class DataType
 			"Any",
 			"Boolean",
 			"Integer",
-			"Double",
+			"Float",
 			"Decimal",
 			"String",
 			"DateTime",
